+++ tile_engine/ops/gemm/gemm_instance_builder.py
@@ -21,7 +21,8 @@ def _import_validation_utils():
 
     # Load the module dynamically
     spec = importlib.util.spec_from_file_location(
-        "validation_utils", os.path.join(parent_dir, "commons", "validation_utils.py")
+        "validation_utils",
+        os.path.join(parent_dir, "commons", "gemm_validation_utils.py"),
     )
     validation_utils = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(validation_utils)
